+++ climpred/alignment.py
@@ -59,7 +59,7 @@ def return_inits_and_verif_dates(forecast, verif, alignment, reference=None, his
     if "valid_time" not in forecast.coords:  # old: create init_lead_matrix
         init_lead_matrix = _construct_init_lead_matrix(forecast, n, freq, leads)
     else:  # new: use valid_time(init, lead)
-        init_lead_matrix = forecast["valid_time"].drop("valid_time").rename(None)
+        init_lead_matrix = forecast["valid_time"].drop_vars("valid_time").rename(None)
     if dask.is_dask_collection(init_lead_matrix):
         init_lead_matrix = init_lead_matrix.compute()