+++ climpred/options.py
@@ -2,7 +2,7 @@ from .constants import GROUPBY_SEASONALITIES
 
 OPTIONS = {
     "seasonality": "month",
-    "perfect_model_persistence_from_initialized_lead_0": False,
+    "PerfectModel_persistence_from_initialized_lead_0": False,
     "warn_for_failed_PredictionEnsemble_xr_call": True,
     "warn_for_rename_to_climpred_dims": True,
     "warn_for_init_coords_int_to_annual": True,
@@ -13,7 +13,7 @@ _SEASONALITY_OPTIONS = frozenset(GROUPBY_SEASONALITIES)
 
 _VALIDATORS = {
     "seasonality": _SEASONALITY_OPTIONS.__contains__,
-    "perfect_model_persistence_from_initialized_lead_0": lambda choice: choice
+    "PerfectModel_persistence_from_initialized_lead_0": lambda choice: choice
     in [True, False, "default"],
     "warn_for_PredictionEnsemble_xr_call": lambda choice: choice
     in [True, False, "default"],
@@ -37,8 +37,8 @@ class set_options:
             :py:meth:`~climpred.classes.HindcastEnsemble.remove_bias`.
         - Allowed: [``"dayofyear"``, ``"weekofyear"``, ``"month"``, ``"season"``]
         - Default: ``dayofyear``.
-    - ``perfect_model_persistence_from_initialized_lead_0``
-        - Use ``climpred.reference.compute_persistence_from_first_lead`` in PerfectModelEnsemble.verify/bootstrap(reference='persistence'). If False uses ``climpred.reference.compute_persistence``.
+    - ``PerfectModel_persistence_from_initialized_lead_0``
+        - Use ``climpred.reference.compute_persistence_from_first_lead`` in PerfectModelEnsemble.verify/bootstrap(reference='persistence'). If ``False`` uses `climpred.reference.compute_persistence <https://climpred.readthedocs.io/en/stable/api/climpred.reference.compute_persistence.html#climpred.reference.compute_persistence>`_ and if ``True`` uses  `climpred.reference.compute_persistence_from_first_lead <https://climpred.readthedocs.io/en/stable/api/climpred.reference.compute_persistence_from_first_lead.html#climpred.reference.compute_persistence_from_first_lead>`_, see `example <https://climpred.readthedocs.io/en/stable/api/climpred.reference.compute_persistence_from_first_lead.html#climpred.reference.compute_persistence_from_first_lead>`_.
         - Allowed: [True, False]
         - Default: False
     - ``warn_for_failed_PredictionEnsemble_xr_call``